+++ features/support/env.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+require 'byebug'
+require 'capybara'
+require 'capybara/cucumber'
+require 'cucumber'
+require 'fig_newton'
+require 'selenium-webdriver'
+require 'webdrivers'
+require_relative 'test_helper'
+
+Fig = FigNewton
+Fig.yml_directory = 'fixtures'
+
+TestHelper.remove_report_dir(FigNewton.reports_dir)
+system("REPORT_PATH=#{FigNewton.test_report_path} bundle exec rspec --format RspecPrettyHtmlReporter spec")