+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.1.4] - 2022-07-22
+
+- Added a row for totals to the Report Overview table
+- Switched to using Cucumber for integration tests
+
 ## [1.1.3] - 2022-04-12
 
 - Reports can generated using a selected theme